requirements.txt
langchain==0.3.27
langchain_community==0.3.27
langchain_core==0.3.72
langchain_openai==0.3.28
python-dotenv==1.1.1
